Divide emulator and parser into different files
